BUILD: buffers: buf.h requires unistd to get ssize_t on libmusl
Building with musl and gcc-5.3 for MIPS returns this :

include/common/buf.h: In function 'b_dist':
include/common/buf.h:252:2: error: unknown type name 'ssize_t'
  ssize_t dist = to - from;
  ^
Including stdint or stddef is not sufficient there to get ssize_t,
unistd is needed as well. It's likely that other platforms will have
the same issue. This patch also addresses it in ist.h and memory.h.

+++ include/proto/htx.h
@@ -440,10 +440,10 @@ static inline void htx_dump(struct htx *htx)
 {
         int32_t pos;
 
-        fprintf(stderr, "htx:%p [ size=%u - data=%u - used=%u - wrap=%s - extra=%lu]\n",
+        fprintf(stderr, "htx:%p [ size=%u - data=%u - used=%u - wrap=%s - extra=%llu]\n",
                 htx, htx->size, htx->data, htx->used,
                 (!htx->used || htx->tail+1 == htx->wrap) ? "NO" : "YES",
-		(unsigned long)htx->extra);
+		(unsigned long long)htx->extra);
         fprintf(stderr, "\thead=%d - tail=%u - front=%u - wrap=%u\n",
                 htx_get_head(htx), htx->tail, htx->front, htx->wrap);
